+++ R/xportr-package.R
@@ -117,7 +117,6 @@
 #'   cli_alert_danger cli_warn start_app stop_app
 #' @importFrom tidyselect all_of any_of where
 #' @importFrom utils capture.output str tail packageVersion
-#' @importFrom stringr str_detect str_extract str_replace str_replace_all str_sub
 #' @importFrom readr parse_number
 #' @importFrom purrr map_chr map2_chr walk iwalk map map_dbl pluck map_lgl
 #' @importFrom graphics stem